requirements.txt
langchain==0.3.4
langchain-cohere==0.3.1
langchain-core==0.3.13
langgraph==0.2.39
langgraph-checkpoint-sqlite==2.0.1
pinecone-client==5.0.1
python-dotenv==1.0.1
requests==2.32.3